Move stuff in/to src/Elewant/Reporting/DomainModel
